CMake: make sure to use the OpenCV include directories
Up until now, this kind of worked because the OpenCV versions that
we installed in /opt used same headers layout as the system-provided
version; we implictly ended up using the latter's headers, and we
got away with it.

However, when compiling with OpenCV 3.0.0 installed in /opt, using
the 2.4.X headers located in /usr results in errors.