Toolbox: Calibration wizard: initial implementation of live-capture pattern detection
The idea here is to allow user to manually trigger detection in
the images that are currently available in the active image
pair source. This way, they can immediately see which images are
good for pattern detection and which not (in contrast to capturing
the image pairs from Image Pair Source dialog and later loading
them into Calibration Wizard).